% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{above_height_surface}
\alias{above_height_surface}
\title{Obtain function of surface above a given height}
\usage{
above_height_surface(profile_equation, hbc, tree_height = 100)
}
\arguments{
\item{profile_equation}{a one-dimensional function that returns the radius of
the tree for a given height}

\item{hbc}{height at the base of the crown}

\item{tree_height}{Tree height}
}
\value{
a function that returns the surface of the tree crown
above a certain height
}
\description{
Obtain function of surface above a given height from a
profile function locally parametrized for a tree and this tree height.
}
\examples{
profile_function <- function(x) {
if(x<3){return(0)}
if(x>5){return(0)}
return(1)
}
surface_function <- above_height_surface(profile_function, 3, 5)
surface_function(0)
surface_function(4)
}
