% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{above_height_volume}
\alias{above_height_volume}
\title{Obtain function of volume above a given height}
\usage{
above_height_volume(profile_equation, tree_height = 100)
}
\arguments{
\item{profile_equation}{a one-dimensional function that returns the radius of
the tree for a given height}

\item{tree_height}{Tree height}
}
\value{
a function that returns the volume of the tree crown
above a certain height
}
\description{
Obtain function of volume above a given height from a
profile function locally parametrized for a tree and this tree height.
}
\examples{
profile_function <- function(x) {
if(x<3){return(0)}
if(x>5){return(0)}
return(1)
}
volume_function <- above_height_volume(profile_function, 5)
volume_function(0) # should be 2*pi
volume_function(4) # should be pi
}
