% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_kernels.R
\name{alemdag}
\alias{alemdag}
\title{Alemdag growing space index}
\usage{
alemdag(imarks, jmarks, dists, dranks, par = list(dbh_mark = 1))
}
\arguments{
\item{imarks}{Marks for the subject plant, a 1-row data frame.}

\item{jmarks}{Data frame with marks for competitors}

\item{dists}{Vector of distances between the subject plant and the
competitors.}

\item{dranks}{Distance ranks.}

\item{par}{List of parameters.}
}
\value{
Numeric vector of length equal to the length of dists containing the
value of the effect of the competition of each the trees j over the tree i.
}
\description{
Competition index using \insertCite{alemdag;textual}{sepci} potentially
available growing space measure to build and index as described in
\insertCite{dobbertin1992;textual}{sepci}
}
\details{
The values of par are taken from the argument kerpar of pairwise(),
if not NULL.

The marks are a data frame, dbh_mark must be the number or name of the column
with the plant dbh.

The effect of each tree j is computed following the formula:
\deqn{\pi \left(\frac{dist_{ij}*tree_idbh}{tree_idbh+tree_jdbh}\right)
^2\frac{\frac{tree_jdbh}
{dist_{ij}}}{\sum \frac{tree_jdbh}{dist_{ij}}}}
}
\examples{
library(siplab)
me <- pairwise(valdepoza, maxR=10, kernel=alemdag,
kerpar = list(dbh_mark = 1))
}
\references{
\insertAllCited{}
}
