% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_kernels.R
\name{bella}
\alias{bella}
\title{Bella influence zone overlap index}
\usage{
bella(
  imarks,
  jmarks,
  dists,
  dranks,
  par = list(dbh_mark = 1, largestCrownRadiusMark = 2)
)
}
\arguments{
\item{imarks}{Marks for the subject plant, a 1-row data frame.}

\item{jmarks}{Data frame with marks for competitors}

\item{dists}{Vector of distances between the subject plant and the
competitors.}

\item{dranks}{Distance ranks.}

\item{par}{List of parameters.}
}
\value{
Numeric vector of length equal to the length of dists containing the
value of the effect of the competition of each the trees j over the tree i.
}
\description{
#' @description
Competition index kernel for the siplab package \insertCite{siplab}{sepci}
implementing \insertCite{bella1971;textual}{sepci}
competition index.
}
\details{
The values of par are taken from the argument kerpar of pairwise(),
if not NULL.

The marks are a data frame, dbh_mark and largestCrownRadiusMark must be the
number or name of the column with the plant dbh and the largest crown radius
respectively. The areas of the influence zones are computed as circle areas.

The effect of each tree j is computed following the formula:
\deqn{\frac{tree_jdbh* AreaOverlapInfluenceZones}
{tree_idbh * AreaInfluenceZoneTree_i}}
}
\examples{
library(siplab)
me <- pairwise(valdepoza, maxR=10, kernel=bella,
kerpar = list(dbh_mark = 1,largestCrownRadiusMark=5))
}
\references{
\insertAllCited{}
}
