% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional_functions.R
\name{ccs_cparea_cone}
\alias{ccs_cparea_cone}
\title{Crown cross sectional crown projection area using a cone}
\usage{
ccs_cparea_cone(
  height_beginning_crown,
  maximum_crown_width_radius,
  tree_height,
  height_cross_sectional_area
)
}
\arguments{
\item{height_beginning_crown}{Height were the crown begins.}

\item{maximum_crown_width_radius}{Radius of the maximum width of the crown.}

\item{tree_height}{Height of the tree.}

\item{height_cross_sectional_area}{Height were we want to get the cross
sectional area}
}
\value{
The cross sectional crown projection area of the crown at
height_cross_sectional_area.
}
\description{
Crown cross sectional crown projection area using a cone
}
\examples{
ccs_cparea_cone(3,2,5,3.5)
}
