% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional_functions.R
\name{ccs_cparea_ellipsoid}
\alias{ccs_cparea_ellipsoid}
\title{Crown cross sectional crown projection area using two ellipsoid}
\usage{
ccs_cparea_ellipsoid(
  height_beginning_crown,
  maximum_crown_width_height,
  maximum_crown_width_radius,
  tree_height,
  height_cross_sectional_area
)
}
\arguments{
\item{height_beginning_crown}{Height were the crown begins.}

\item{maximum_crown_width_height}{Height were the crown reaches its maximum
width,}

\item{maximum_crown_width_radius}{Radius of the maximum width of the crown.}

\item{tree_height}{Height of the tree.}

\item{height_cross_sectional_area}{Height were we want to get the cross
sectional area}
}
\value{
The cross sectional area of the crown at height_cross_sectional_area
if height_cross_sectional_area is in the lower ellipsoid it will return
the area at maximum_crown_width_height
}
\description{
Crown cross sectional crown projection area using two ellipsoid
}
\examples{
ccs_cparea_ellipsoid(3,4,2,5,3.5)
}
