% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional_functions.R
\name{ccs_volume_ellipsoid}
\alias{ccs_volume_ellipsoid}
\title{Crown cross sectional volume using two ellipsoid}
\usage{
ccs_volume_ellipsoid(
  height_beginning_crown,
  maximum_crown_width_height,
  maximum_crown_width_radius,
  tree_height,
  height_cross_sectional_v
)
}
\arguments{
\item{height_beginning_crown}{Height were the crown begins.}

\item{maximum_crown_width_height}{Height were the crown reaches its maximum
width,}

\item{maximum_crown_width_radius}{Radius of the maximum width of the crown.}

\item{tree_height}{Height of the tree.}

\item{height_cross_sectional_v}{Height were we want to get the cross
sectional volume}
}
\value{
The cross sectional volume of the crown above
height_cross_sectional_v.
}
\description{
Crown cross sectional volume using two ellipsoid
}
\examples{
ccs_volume_ellipsoid(3,4,2,5,3.5)
}
