% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cparea_at_height}
\alias{cparea_at_height}
\title{Obtain function of crown projection area at a given height}
\usage{
cparea_at_height(profile_equation, hlcr)
}
\arguments{
\item{profile_equation}{a one-dimensional function that returns the radius of
the tree for a given height}

\item{hlcr}{Height of the largest crown radius}
}
\value{
a function that returns the crown projection area of the tree at a
given height
}
\description{
Obtain function of crown projection area at a given height
}
\examples{
profile_function <- function(x) {
if(x<3){return(0)}
if(x>5){return(0)}
return(1)
}
area_function <- cparea_at_height(profile_function, hlcr = 3)
area_function(0) # should be 0
area_function(4) # should be pi
}
