% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_kernels.R
\name{martin_ek}
\alias{martin_ek}
\title{Martin and Ek exponential weighting of relative diameters index}
\usage{
martin_ek(imarks, jmarks, dists, dranks, par = list(dbh_mark = 1))
}
\arguments{
\item{imarks}{Marks for the subject plant, a 1-row data frame.}

\item{jmarks}{Data frame with marks for competitors}

\item{dists}{Vector of distances between the subject plant and the
competitors.}

\item{dranks}{Distance ranks.}

\item{par}{List of parameters.}
}
\value{
Numeric vector of length equal to the length of dists containing the
value of the effect of the competition of each the trees j over the tree i.
}
\description{
Competition index kernel for the siplab package \insertCite{siplab}{sepci}
implementing \insertCite{martinek;textual}{sepci}
competition index.
}
\details{
The values of par are taken from the argument kerpar of pairwise(),
if not NULL.

dbh_mark in par must be 1 if there is only one mark. If the marks are
a data frame, dbh_mark must be the number or name of the column with the
plant dbh.

The effect of each tree j is computed following the formula:
\deqn{\frac{tree_jdbh}{tree_idbh}e^{16
\frac{distance(tree_i,tree_j)}{tree_idbh+tree_jdbh}}}
}
\examples{
library(siplab)
me <- pairwise(valdepoza, maxR=10, kernel=martin_ek,
kerpar = list(dbh_mark = 1))
}
\references{
\insertAllCited{}
}
